Colour vertices of the icosahedron
The rainbow array was meant to colour the 10 quads
that can be made out of an icosahedron; not the vertices.
So it just looks like delicious rainbow soup right now.

Committing as-is because it looks kinda cool, and because
this represents my first deliberate change to a shader!
(Hurrah!) I'll probably need to actually sit down and read
some GLSL documentation at some point rather than just
guessing at how it fits together... ;)