Move to having 5 root quads instead of 10
Using 10 was a misguided. I had wanted to keep the root quads square
for the sake of simplicity, but in retrospect that actually makes things
more complicated.

This slightly complicates the `project` method, but it means that when
it comes time to do all calculus on positions and vectors in voxel space
there will only be one kind of root quad (as opposed to different logic
for northern hemisphere quads vs. southern hemisphere), and logic about
propagating edge voxels between chunks will be simpler.

This revealed a tiny bug where I had used the x-resolution and y-resolution
the wrong way. Having a deliberately non-square world resolution will now
very quickly reveal any similar bugs, so there's a bonus.