Copy cells from chunks that own them
Copy cells over from chunks that own cells to those that
contain the same cells but don't own them.

This resolves the last obvious visual glitch: hard lines down
the middle of rows of cells where the chunk on each side had
a different idea of what is in that cell. Now both chunks will
end up with a copy of the authoritative state. Hurrah! :)