Send Encoders between render sys and main
This is heavily inspired by the way Yasteroids does it,
but is kind of silly at the moment because I'm calling the
rendering "system" synchronously still. For all this to be
at all worthwhile I need to turn my rendering "system" into
and actual Specs system so it can run asynchronously.