Add dummy visual for cell dweller
This currently sits at the middle of the world. Next up is to position
it where the cell dweller actually is.

This also reminds me that my example globe is stupidly small;
I should scale it up, scale up my camera movement speed, and make
the dummy mesh default to a radius of ~1.