Implement turning, including on pentagons
This is starting to reveal to me that the logic in `maybe_rebase_on_adjacent_root`,
although working for now, might actually be structured poorly to handle the final
forward movement case, which is moving through pentagons.

Consider moving from (0, 0) in the positive y-axis. In the first pentagon
you meet, naively ploughing ahead with rebase code that doesn't have
any idea where you _came from_ will happily assume we came from triangle 6,
and do nothing at all. So the simple approach of "move or rotate,
and then separately rebase" is not actually valid.