WIP chunk view creation system
We're temporarily doing useless double-work in mesh
generation, and the new stuff is just being thrown away.

This is because I'm currently trying to figure out the
best way to create the OpenGL resources (VBO etc.).
The obvious way to do it would be to either create a new
resource factory in the render system, or pass a cloned
one in, but this isn't possible; they're not `Send`
(at least for OpenGL gfx backend).

I'm leaning toward manually iterating over all Visuals
on the main thread where I already have access to a
resource factory, and realising `ProtoMesh`es there.