Finish chunk view creation system
By "finish" I mean switch over to using it for all chunk geometry,
and remove the original code.

There is however still a lot of clean-up work to come. For example,
to get around the OpenGL resource factory not being `Send`, I've exposed
the collection of `Mesh`es on the render system through an `Arc`,
and I'm then realising proto-meshes in the main/device thread.
If this turns out to be the general shape of my long-term solution,
then I'll at very least need to factor out a `MeshRepository` or similar
with a sensible interface, instead of just exposing the render system's
privates in an the ugly and error-prone way I've just done.

Consider this hacking in the general direction of a solution.