Take multiple walks in movement tests
One of them was intermittently failing. It turned out to be that
I was forgetting to switch the turn bias if the end of the walk
happened to land on a pentagon.

Backward stepping has this built in (to the non-test code) so that
is why only one of these tests was failing.