Add mining system
It's really not obvious that this works at all right now,
because the chunk views aren't updated. But if you press 'U'
when looking at a block of dirt/grass, then you'll be able to
walk through where it used to be.

Next up is to dynamically re-generate the meshes.