Better face culling, but too slow
I knew this approach would be slow, but I hadn't realised
just how bad it would be.

This is sufficiently slow that I'm not even comfortable
pushing it to master.

The next step will be to write a benchmark test so I can
compare this against attempts at a faster implementation.