Fix problem of chunk views not always updating
I've done this in the most disgusting and slow way imaginable.

But aside from the slowness, the hacks are contained, and I want
to get this branch merged back into master, so I'm going to leave
it as is and then revisit it later as I build out better internal
APIs to support this kind of thing. Specifically, I need a function
to tell me all the chunks that include a given cell.