Create chunk system
Factored out existing chunk creation logic into a system
which allows Globe to be a bit dumber.

This system will soon be smarter about what chunks to create when,
but for now I've just moved the existing logic into it and shuffled
things around as necessary to make that work.