Limit how many chunks can be loaded at a time
This unloads chunks, but doesn't clean up meshes for them.
So eventually we will fill up the video card with garbage VBOs.

I'm being lazy and not fixing this for now because it doesn't show
up in the benchmark tests, on account of them not actually setting
up the graphics subsystem.