Move a bunch of logic back into Globe
It's a bit gross having it know/do so much, but it was currently more
gross to have to find the ChunkSystem just to find the surface of the globe.

I'm thinking about pushing the dumb bits of the globe voxmap further down
into their own struct. That might be what I actually want. When it comes time
to load and save chunks in addition to creating them, I imagine that a lot of
this will become clearer.