Fix noise-rs abuse
The API changed. I was constantly recreating the permutation table,
because it's now inside the different noise generator structs.

Unfortunately enough has changed in noise-rs that the terrain now
looks somewhat different. It's not that I was attached at all to the
way it was before, but now I understand the nature of the noise less. :)