Add CursorMut, and use it in new globe_ext module
Introduced CursorMut so that it can ensure the presence of any chunks
it needs to work with.

This comes with (what I assume to be) a significant performance regression
for (non-mutable) Cursor, because I'm now only storing the chunk origin in
each kind of cursor instead of a reference to the chunk, which we can't have
in the mutable cursor. This means we have to look the chunk up each time we
get a cell reference.

I haven't bothered comparing benchmarks to see how much worse this is at the
moment, because I plan to specialise the cursors to reintroduce the old behavior
for non-mutable cursors ASAP.