Bail out ASAP when building neighbour chunks
Don't do a bunch of redundant checks if there's no way the current
cell is relevant. I haven't actually tested that this speeds things up,
but I'm planning to reimplement this soon in a much more efficient way anyway,
so I'm just going to make a wild assumption for now.