Add helper to find chunks for a point
Finds all chunks in the current root that contain the given point.

I'm intending to write another iterator for all the different representations
of a point (i.e. in all the roots it can be expressed in) and then combining
the two to find all the chunks and points within them we need to copy values into when
changing the content of a cell in its owning chunk.

This will let me finally get rid of `copy_all_authoritative_cells` for good!