Add AutoResource trait
Used for Resources that know how to construct
themselves and ensure that there is an instance
of them present in a given Specs World.

This is handy for untangling dependencies between
resources and systems; having a helper like this should
encourage me to put more in Resources wherever possible,
rather than in System state.