Add web (Emscripten) demo
Tries to create a simple world and run it.
Currently fails building in some late stage of Emscripten.
After I get it building, I'll need to patch some of the crates
I'm using to allow using WebGL — unless they get there first.

I've previously had most of PlanetKit working, minus the rendering,
and hacked that in (badly) with Three.js.

See https://github.com/jeffparsons/planetkit/pull/11 for that attempt.

I'm not interested in pursuing that as a long-term approach, so I'm
just going to merge the small changes that don't break anything else
to prevent them from rotting on a branch forever.