Add web (Emscripten) demo
Currently just creates a globe and prints its radius.

I'll probably wait until the gfx-ll work stabilises to try getting
rendering from Rust in a browser working, but I might be able to
get vertex data out and shove it into Three.js for a neat demo
in the interim.