Show chunks using Three.js
This is a major hack, but it gets some pixels into the browser.
They're even the right colours!

Apparently gfx-rs "pre-ll" does actually build on Emscripten now,
so I'm going to give that a shot. If it works for me, then I'll rip
out all the Three.js stuff and build a proper web front-end for
PlanetKit, and merge the whole thing.