Create Glutin window manually
Rather than letting PistonWindow create it.
This gives us a little more control over window initialisation,
which I think we're going to need for getting WebGL working.

This doesn't actually work in the browser yet, but it can't
hurt to have this stuff on master to make my experimental branches
deviate a bit less from master.