Use nphysics to handle grenade/ground collisions
This makes the current leakiness around chunk views even worse for now,
because we're now also adding chunk geometry to the physics world every
time we generate a chunk view, and then never cleaning it up.

I intend to generalise some of the nphysics integration bits (move it out
of Kaboom and into PlanetKit) and start cleaning up chunk views (etc.)
properly when chunks are removed from the world.