Clean up colliders for chunk physics meshes
Whenever a chunk view is removed, remove any existing
physics mesh for it.

Whenever a chunk view is regenerated, remove any existing
physics mesh for it, then create a new one.

A lot of this is really unpleasant and easy to stuff up.
I need to put some thought into making this ergonomic.