Rename GridPoint3 -> Point3
It's in the 'planetkit_grid' crate now, and I'm pretty well
convinced that it's better to always just use 'grid' and then
explicitly qualify it as `grid::Point3` to avoid confusion with
the Nalgebra point types, which will quite often be used in the
same context.

Next commit: same thing for GridPoint2.