Fix a bunch of Clippy lints
The only remaining errors are about type complexity
and number of function arguments. This is all related
to my use of Specs storage types... and I'm not sure if there's
a much better way to express what I'm doing.

I'll put some thought into that, but it'll probably end up being
a bunch of `allow`s to stop Clippy complaining.